CodeForce Solve
CodeForce Solve
C2Ladder 
Problem : Same Differences
Difficulty : 1200
complexity : 296 ms || 7300 KB

+++ 1520D.cpp
@@ -10,14 +10,14 @@ int main(){
     
     while(t != 0){
 
-        long long n;
+        int n;
         cin>>n;
-        vector<long long> a(n);
+        vector<int> a(n);
         for (int i = 0; i < n; i++){
             cin>>a[i];
         }
 
-        unordered_map<long long, long long> map;
+        unordered_map<int, int> map;
         long long count = 0;
         for(int i = 0; i < n; i++){
             int key = a[i] - i;
